Fix seed ratio by removing seedOverride check
When a torrent state is changed, the seed values are read from the
per-torrent metadata, if and only if a"seedOverride" option is also set.
If it is not set, the default values are used. There is no way to
actually set this option, however, which means the default values are
always used regardless of what is specified in the config file.

This change removes the seedOverride check, which allows the config
values to actually take effect.