Make service startup report failures to SCM
In cases like issue #212, the service can fail to start for reasons
such as the configured port is already in use. Under the old service
startup model this meant Hadouken would die and the SCM would
eventually time out the start, reporting a generic failure message.

With this change the SCM is contacted earlier in the startup process,
before things really start running, putting the service in a pending
state. Only after Hadouken starts listening for requests does the
service move into the started state. If a failure happens during this
process the service moves from pending to stopping and provides an
error code to the SCM. In all cases the Hadouken service should no
longer time out.